fix(metrics): cast memory values to f64 for Prometheus Gauge::set()
process_resident_memory and process_virtual_memory are Gauge metrics
which expect f64. The kb * 1024 expression produced u64, causing
E0308 mismatched types during release builds.

+++ crates/server/src/metrics.rs
@@ -303,11 +303,11 @@ impl FolioMetrics {
             for line in status.lines() {
                 if line.starts_with("VmRSS:") {
                     if let Some(kb) = extract_kb_value(line) {
-                        self.process_resident_memory.set(kb * 1024);
+                        self.process_resident_memory.set((kb * 1024) as f64);
                     }
                 } else if line.starts_with("VmSize:") {
                     if let Some(kb) = extract_kb_value(line) {
-                        self.process_virtual_memory.set(kb * 1024);
+                        self.process_virtual_memory.set((kb * 1024) as f64);
                     }
                 }
             }
